requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.0
google-auth==2.26.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
orjson==3.9.10
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-core==2.14.5
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
PyJWT==2.8.0
pymongo==4.6.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tqdm==4.66.1
typing-extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
